Remove "(or more)" in Site-Declared Sets in Browsers section
rhiaro called out when reviewing https://github.com/privacycg/first-party-sets/pull/45 that the mention of "**All sites** with use-cases that depend on cross-domain, same-party communication will be required to declare one (or more) sets." is confusing assuming "site" corresponds to the HTML definition of a "tuple of URL scheme, and domain". The section later goes on to describe how a domain cannot be part of more than one set; so this is contradictory.

The original intention for the phrase was to say that organizations can define more than one First-Party Set; but a single domain itself can only appear in one set.