Add page numbers for FTC report
The entire 112-page report makes 74 references to "party" without
limiting "party" to an individual legal entity. (This may because
many well-known online brands are actually multiple entities.)

The report also mentions that the same company may be different
"parties" if one service is visible to the user while the other one
is not.