Introduced a requirement for domains in the First-Party Set to provide cryptographic proof at a well-known end point that they adhere to the common privacy policy.
Therefore removed reference to brand throughout the document as the verifiable common privacy policy provides all the proof needed. Brand will be very hard for an enforcement entity to rule on as it is subjective.

Removed the need for the website operator to display information about the First Party Set as this will now be possible for the User Agent to perform via the cryptographic proof provided at the well-known end point by each domain in the set. This will also provide a consistent user interface across all First-Party Sets and enable the user to find the information in a clear and commonly understood location within the browser. Placing the burden on the website operator will lead to confusion and inconsistency.

Removed optional section on guidance to others from the User Agent as cryptographic proof of common privacy policy no longer requires such guidance.